updated module

+++ db/migration/000001_init_schema.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE accounts (
     id serial NOT NULL,
     owner varchar NOT NULL,
-    balance float NOT NULL,
+    balance bigint NOT NULL,
     currency varchar NOT NULL,
     created_at timestamp NOT NULL DEFAULT (now()),
     PRIMARY KEY (id)
@@ -10,7 +10,7 @@ CREATE TABLE accounts (
 CREATE TABLE entries (
     id serial NOT NULL,
     account_id int NOT NULL,
-    amount float NOT NULL,
+    amount bigint NOT NULL,
     created_at timestamp NOT NULL DEFAULT (now()),
     PRIMARY KEY (id),
     FOREIGN KEY (account_id) REFERENCES accounts(id)
@@ -20,7 +20,7 @@ CREATE TABLE transfers (
     id serial NOT NULL,
     from_account_id int NOT NULL,
     to_account_id int NOT NULL,
-    amount float NOT NULL,
+    amount bigint NOT NULL,
     created_at timestamp NOT NULL DEFAULT (now()),
     PRIMARY KEY (id),
     FOREIGN KEY (from_account_id) REFERENCES accounts(id),
